More Solaris build issues corrected
git-svn-id: svn+ssh://svn.code.sf.net/p/x3270/x3270/trunk@1761 12e4f418-770b-42d6-82dd-bd6eefd7344d

+++ x3270/x3270/keymap.c
@@ -832,7 +832,8 @@ create_text(void)
 	static Widget source = NULL;
 
 	/* Ready a file. */
-	(void) snprintf(km_file, sizeof(km_file), "/tmp/km.%d", getpid());
+	(void) snprintf(km_file, sizeof(km_file), "/tmp/km.%u",
+		(unsigned)getpid());
 	f = fopen(km_file, "w");
 	if (f == NULL) {
 		popup_an_errno(errno, "temporary file open");
